requirements.txt
fastapi
uvicorn[standard]
pyJWT
python-decouple
pydantic[email]